# APIServer/Areas/Metric/Models/User.cs
using APIServer.Utilities;
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Threading.Tasks;

namespace APIServer.Areas.Metric.Models
{
    public class User
    {
        public int Id { get; set; }
    }
}